+++ backend-api/src/functions/asyncCredential/asyncCredentialHandler.test.ts
@@ -5,19 +5,6 @@ import {
 } from "../services/events/tests/mocks";
 import { Logger } from "../services/logging/logger";
 import { MockLoggingAdapter } from "../services/logging/tests/mockLogger";
-import {
-  MockClientRegistryServiceeGetPartialClientInternalServerResult,
-  MockClientRegistryServiceGetPartialClientBadRequestResponse,
-  MockClientRegistryServiceGetPartialClientSuccessResult,
-  MockSessionServiceCreateSessionErrorResult,
-  MockSessionServiceCreateSessionSuccessResult,
-  MockSessionServiceGetActiveSessionSuccessResult,
-  MockSessionServiceGetSessionBySubErrorResult,
-  MockSessionServiceNoActiveSession,
-  MockTokenServiceGetDecodedTokenErrorResult,
-  MockTokenServiceInvalidSignatureErrorResult,
-  MockTokenServiceSuccess,
-} from "../testUtils/asyncCredentialMocks";
 import { MockJWTBuilder } from "../testUtils/mockJwt";
 import { buildRequest } from "../testUtils/mockRequest";
 import { Result, successResult } from "../utils/result";
@@ -29,6 +16,23 @@ import {
   IDecodeToken,
   IVerifyTokenSignature,
 } from "./tokenService/tokenService";
+import {
+  MockTokenServiceGetDecodedTokenErrorResult,
+  MockTokenServiceInvalidSignatureErrorResult,
+  MockTokenServiceSuccess,
+} from "./tokenService/tests/mocks";
+import {
+  MockClientRegistryServiceeGetPartialClientInternalServerResult,
+  MockClientRegistryServiceGetPartialClientBadRequestResponse,
+  MockClientRegistryServiceGetPartialClientSuccessResult,
+} from "../services/clientRegistryService/tests/mocks";
+import {
+  MockSessionServiceCreateSessionErrorResult,
+  MockSessionServiceCreateSessionSuccessResult,
+  MockSessionServiceGetActiveSessionSuccessResult,
+  MockSessionServiceGetSessionBySubErrorResult,
+  MockSessionServiceNoActiveSession,
+} from "./sessionService/tests/mocks";
 
 const env = {
   SIGNING_KEY_ID: "mockKid",
