+++ backend-api/src/functions/asyncCredential/sessionService/tests/mocks.ts
@@ -1,0 +1,107 @@
+import { errorResult, Result, successResult } from "../../../utils/result";
+import { IGetActiveSession, ICreateSession } from "../sessionService";
+
+export class MockSessionServiceGetSessionBySubErrorResult
+  implements IGetActiveSession, ICreateSession
+{
+  readonly tableName: string;
+  readonly indexName: string;
+
+  constructor(tableName: string, indexName: string) {
+    this.tableName = tableName;
+    this.indexName = indexName;
+  }
+
+  getActiveSession = async (): Promise<Result<string | null>> => {
+    return errorResult({
+      errorMessage: "Mock failing DB call",
+      errorCategory: "SERVER_ERROR",
+    });
+  };
+
+  createSession = async (): Promise<Result<string>> => {
+    return successResult("mockSessionId");
+  };
+}
+
+export class MockSessionServiceNoActiveSession
+  implements IGetActiveSession, ICreateSession
+{
+  readonly tableName: string;
+  readonly indexName: string;
+
+  constructor(tableName: string, indexName: string) {
+    this.tableName = tableName;
+    this.indexName = indexName;
+  }
+  getActiveSession = async (): Promise<Result<string | null>> => {
+    return successResult(null);
+  };
+
+  createSession = async (): Promise<Result<string>> => {
+    return successResult("mockSessionId");
+  };
+}
+
+export class MockSessionServiceGetActiveSessionSuccessResult
+  implements IGetActiveSession, ICreateSession
+{
+  readonly tableName: string;
+  readonly indexName: string;
+
+  constructor(tableName: string, indexName: string) {
+    this.tableName = tableName;
+    this.indexName = indexName;
+  }
+
+  getActiveSession = async (): Promise<Result<string | null>> => {
+    return successResult("mockSessionId");
+  };
+
+  createSession = async (): Promise<Result<string>> => {
+    return successResult("mockSessionId");
+  };
+}
+
+export class MockSessionServiceCreateSessionErrorResult
+  implements IGetActiveSession, ICreateSession
+{
+  readonly tableName: string;
+  readonly indexName: string;
+
+  constructor(tableName: string, indexName: string) {
+    this.tableName = tableName;
+    this.indexName = indexName;
+  }
+
+  getActiveSession = async (): Promise<Result<string | null>> => {
+    return successResult(null);
+  };
+
+  createSession = async (): Promise<Result<string>> => {
+    return errorResult({
+      errorMessage: "Mock error",
+      errorCategory: "SERVER_ERROR",
+    });
+  };
+}
+
+export class MockSessionServiceCreateSessionSuccessResult
+  implements IGetActiveSession, ICreateSession
+{
+  readonly tableName: string;
+  readonly indexName: string;
+
+  constructor(tableName: string, indexName: string) {
+    this.tableName = tableName;
+    this.indexName = indexName;
+  }
+
+  getActiveSession = async (): Promise<Result<string | null>> => {
+    return successResult(null);
+  };
+
+  createSession = async (): Promise<Result<string>> => {
+    return successResult("mockSessionId");
+  };
+}
