+++ backend-api/src/functions/asyncToken/tokenService/tests/mock.ts
@@ -1,0 +1,17 @@
+import { successResult, errorResult, Result } from "../../../utils/result";
+import { IMintToken } from "../tokenService";
+
+export class MockTokenServiceSuccessResult implements IMintToken {
+  async mintToken(): Promise<Result<string>> {
+    return successResult("mockToken");
+  }
+}
+
+export class MockTokenServiceErrorResult implements IMintToken {
+  async mintToken(): Promise<Result<string>> {
+    return errorResult({
+      errorMessage: "Failed to sign Jwt",
+      errorCategory: "SERVER_ERROR",
+    });
+  }
+}
