+++ backend-api/src/functions/services/clientRegistryService/tests/clientRegistryService.test.ts
