+++ backend-api/src/functions/tests/pact/createApp.ts
@@ -1,10 +1,10 @@
 import express, { Application, Request, Response } from "express";
-import { asyncTokenConfig } from "./asyncToken/asyncTokenConfiguration";
-import { asyncCredentialConfig } from "./asyncCredential/asyncCredentialConfiguration";
 import { buildLambdaContext } from "../../testUtils/mockContext";
 import { buildRequest } from "../../testUtils/mockRequest";
 import { lambdaHandlerConstructor as tokenLambdaHandlerConstructor } from "../../asyncToken/asyncTokenHandler";
 import { lambdaHandlerConstructor as credentialLambdaHandlerConstructor } from "../../asyncCredential/asyncCredentialHandler";
+import { asyncTokenDependencies } from "./dependencies/asyncTokenDependencies";
+import { asyncCredentialConfig } from "./dependencies/asyncCredentialDependencies";
 
 export async function createApp(): Promise<Application> {
   const app = express();
@@ -14,7 +14,7 @@ export async function createApp(): Promise<Application> {
 
   app.post("/async/token", async (req: Request, res: Response) => {
     const result = await tokenLambdaHandlerConstructor(
-      asyncTokenConfig.dependencies,
+      asyncTokenDependencies.dependencies,
       buildLambdaContext(),
       buildRequest({ headers: req.headers, body: req.body }),
     );
